chore: Add 400.html template for invalid requests
This commit adds a new template, 400.html, to handle invalid requests. The template extends the base.html template and displays a custom error message for a 400 status code. It also includes a link to return to the dashboard. This template will be used when a client makes a request with invalid syntax.